Move hook files to components dir, get text for blobs
